moved styles from main-page component to its child components
