+++ src/components/left-section/left-section.component.js
