+++ src/components/top-section/top-section.component.js
@@ -1,0 +1,13 @@
+import React from "react";
+
+
+const TopSection = () => {
+    return (
+        <div className="text-center" style={{backgroundColor: "blanchedalmond", height:"100%"}}>
+            <h1>AppetAIzing</h1>
+        </div>
+        
+    )
+}
+
+export default TopSection;
