--- src/components/top-section/top-section.js
@@ -1,18 +1,0 @@
-import React from "react";
-import { Container, Row, Col } from "react-bootstrap";
-import "./top-section.styles.css";
-
-const TopSection = () => {
-    return (
-        // <Container className="d-flex justify-content-center" fluid style={{backgroundColor: "blanchedalmond", height: "10%"}}>
-        //     <Row className="d-flex">
-        //         <Col className="d-flex align-items-center">
-        //         <h1 className="d-flex">AppetAIzing</h1>
-        //         </Col>
-        //     </Row>
-        // </Container>
-        <h1>AppetAIzing</h1>
-    )
-}
-
-export default TopSection;
