--- src/components/top-section/top-section.styles.css
@@ -1,4 +1,0 @@
-/* .top-section {
-    height: 10vw;
-    background-color: blanchedalmond;
-} */
